Move "App.tsx" to "app/index.tsx"
