+++ packages/playground/__tests__/Welcome.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
-import { Welcome } from '~/app';
+import { Welcome } from '~/app/WasmTest/Welcome';
 
 test('should render the "Welcome" component', () => {
   render(<Welcome />);
-  const linkElement = screen.getByText(/Welcome to the ink! Playground!/i);
+  const linkElement = screen.getByText('Welcome to the ink! Playground!');
   expect(linkElement).toBeInTheDocument();
 });
