+++ packages/playground/src/app/index.tsx
@@ -1,5 +1,20 @@
-export { Console } from './Console';
-export { Editor } from './Editor';
-export { Layout } from './Layout';
-export { Header } from './Header';
-export { Welcome } from './WasmTest/Welcome';
+import { Console } from './Console';
+import { Editor } from './Editor';
+import { Layout } from './Layout';
+import { Header } from './Header';
+import { AppProvider } from '~/context';
+import { ReactElement } from 'react';
+
+const App = (): ReactElement => {
+  return (
+    <AppProvider>
+      <Layout>
+        <Header />
+        <Editor />
+        <Console />
+      </Layout>
+    </AppProvider>
+  );
+};
+
+export default App;
