+++ packages/playground/src/style/tailwind.css
@@ -13,9 +13,3 @@
 .verticalDivider {
   @apply border-l max-h-8 mt-4 dark:border-dark border-light;
 }
-
-/* temporary */
-
-.btn {
-  @apply bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded mx-2;
-}
